requirements.txt
matplotlib~=3.3.4
Pillow~=8.4.0
imageio~=2.13.4
numpy~=1.19.5
ipython~=7.16.2
tensorflow-gpu~=2.6.2
tensorflow_addons~=0.14.0